fix: restore red master + dependency maintenance & documentation pass (v1.53.2)
master had been red across every merge-gating workflow since v1.53.1 merged on
2026-07-19: Tests, Build Windows, Build Android and the scheduled Build iOS all
failed to compile.

Build fix
- commands/subcultures.rs passed passage_number (an i32, from subculture_count)
  to signed_ledger::lifecycle::passage, which takes an i64 — introduced when
  WP-75 wired signed lifecycle events into the passage path. Widened with
  i64::from(...).
- The code lives behind the `tauri-commands` feature, so the fast local gates
  (--no-default-features test + clippy, npm test, npm run check) never compile
  it. Verified here against the real CI gate: 677 Rust tests and clippy clean
  with full features.

Dependencies (lockfile only, package.json untouched)
- npm audit fix closed two high-severity advisories: fast-uri host confusion
  (GHSA-v2hh-gcrm-f6hx) and postcss path traversal (GHSA-r28c-9q8g-f849).
- Brought four in-range-drifted packages current.
- package-lock.json's version field still read 1.48.0, five releases stale.
- Still open and disclosed: xlsx (no upstream fix, reachable from Excel import)
  and a dev-only brace-expansion copy nested under vite-plugin-pwa, whose only
  offered remedy is a downgrade — not applied.

Documentation
- ROADMAP.md: replaced a ~7,000-word unbroken header paragraph with a fact
  table plus two collapsible sections; per-migration history is now a 52-row
  table. No fact dropped.
- UserManual.md: was stale at v1.45.0 with no Phase G registry/coordination and
  no Phase H content at all. New header, linked TOC, and six new sections
  (27-32) covering on-chain anchoring & the signed ledger, partner-lab
  exchanges, compliance rules & waivers, the submission pipeline, the
  data-integrity self-check, and the mycology Fruiting overview.
- docs/README.md: new specification index; uniform header table on all 11 specs.
- skills.md: corrected the next-migration number, added the full-feature
  verification procedure, the integer-width trap, and a new "Docs that drift"
  section.
- SteloPTC.md, README.md, CHANGELOG.md intro refreshed; every relative link and
  in-document anchor across all 20 Markdown files verified.
- DailyClaudeRoutineCheckup.md rewritten for 2026-07-25.

Verified: cargo test --lib (677) · cargo test --lib --no-default-features (640)
· cargo clippy --lib and --no-default-features, -D warnings (both clean)
· npm test (113) · npm run check (0/0, 418 files) · npm run build.

No schema change (52 migrations, next is 053).

+++ src-tauri/src/commands/subcultures.rs
@@ -308,7 +308,7 @@ pub fn create_subculture(
     // (the WP-67 "sign every mutation" follow-up). Best-effort — a ledger hiccup
     // never fails the passage.
     let (sev_type, sev_payload) =
-        crate::signed_ledger::lifecycle::passage(&request.specimen_id, passage_number, "passage");
+        crate::signed_ledger::lifecycle::passage(&request.specimen_id, i64::from(passage_number), "passage");
     crate::signed_ledger::try_append_signed_event(
         &db.conn, &user.id, sev_type, "specimen", Some(&request.specimen_id), &sev_payload,
     );
